+++ lightyear_inputs_leafwing/src/input_message.rs
@@ -86,7 +86,7 @@ impl<A: LeafwingUserAction> ActionStateSequence for LeafwingSequence<A> {
     /// we have an `ActionState`.
     fn build_from_input_buffer<'w, 's>(
         input_buffer: &InputBuffer<Self::Snapshot, Self::Action>,
-        num_ticks: u16,
+        num_ticks: u32,
         end_tick: Tick,
     ) -> Option<Self> {
         let mut diffs = Vec::new();
